Colorize avatar based on username (#95)
